revert some media factory changes, adjust tests

+++ src/SWP/Bundle/ContentListBundle/Services/ContentListService.php
@@ -75,7 +75,7 @@ final class ContentListService implements ContentListServiceInterface
             ->getQuery()
             ->getResult();
 
-        if (\count($items) > $contentList->getLimit()) {
+        if (null !== $contentList->getLimit() && \count($items) > $contentList->getLimit()) {
             foreach ($items as $key => $item) {
                 if ($key + 1 > $contentList->getLimit()) {
                     $this->contentListItemRepository->remove($item);
